feat: add time based pagination (#124)
This PR adds a mechanism to load events on a per 
minute basis. This means, console queries only for the first
minute (not the whole user selected time range) on the log 
table page. Then as user clicks the load older / newer data
different time ranges are queried.

This way console has to load a much smaller subset of the
overall data.

Co-authored-by: Nitish Tiwari <tiwari.nitish@gmail.com>